+++ clean_python/api_client/files.py
@@ -6,6 +6,7 @@ import re
 from pathlib import Path
 from typing import BinaryIO
 from typing import Callable
+from typing import Dict
 from typing import Optional
 from typing import Tuple
 from typing import Union
@@ -58,6 +59,7 @@ def download_file(
     timeout: Optional[Union[float, urllib3.Timeout]] = 5.0,
     pool: Optional[urllib3.PoolManager] = None,
     callback_func: Optional[Callable[[int, int], None]] = None,
+    headers_factory: Optional[Callable[[], Dict[str, str]]] = None,
 ) -> Tuple[Path, int]:
     """Download a file to a specified path on disk.
 
@@ -75,6 +77,7 @@ def download_file(
             created with a retry policy of 3 retries after 1, 2, 4 seconds.
         callback_func: optional function used to receive: bytes_downloaded, total_bytes
             for example: def callback(bytes_downloaded: int, total_bytes: int) -> None
+        headers_factory: optional function to inject headers
 
     Returns:
         Tuple of file path, total number of downloaded bytes.
@@ -105,6 +108,7 @@ def download_file(
                 timeout=timeout,
                 pool=pool,
                 callback_func=callback_func,
+                headers_factory=headers_factory,
             )
     except Exception:
         # Clean up a partially downloaded file
@@ -124,6 +128,7 @@ def download_fileobj(
     timeout: Optional[Union[float, urllib3.Timeout]] = 5.0,
     pool: Optional[urllib3.PoolManager] = None,
     callback_func: Optional[Callable[[int, int], None]] = None,
+    headers_factory: Optional[Callable[[], Dict[str, str]]] = None,
 ) -> int:
     """Download a url to a file object using multiple requests.
 
@@ -139,6 +144,7 @@ def download_fileobj(
             created with a retry policy of 3 retries after 1, 2, 4 seconds.
         callback_func: optional function used to receive: bytes_downloaded, total_bytes
             for example: def callback(bytes_downloaded: int, total_bytes: int) -> None
+        headers_factory: optional function to inject headers
 
     Returns:
         The total number of downloaded bytes.
@@ -156,6 +162,10 @@ def download_fileobj(
     """
     if pool is None:
         pool = get_pool()
+    if headers_factory is not None:
+        base_headers = headers_factory()
+    else:
+        base_headers = {}
 
     # Our strategy here is to just start downloading chunks while monitoring
     # the Content-Range header to check if we're done. Although we could get
@@ -165,7 +175,7 @@ def download_fileobj(
     while True:
         # download a chunk
         stop = start + chunk_size - 1
-        headers = {"Range": "bytes={}-{}".format(start, stop)}
+        headers = {"Range": "bytes={}-{}".format(start, stop), **base_headers}
 
         response = pool.request(
             "GET",
@@ -210,6 +220,7 @@ def upload_file(
     pool: Optional[urllib3.PoolManager] = None,
     md5: Optional[bytes] = None,
     callback_func: Optional[Callable[[int, int], None]] = None,
+    headers_factory: Optional[Callable[[], Dict[str, str]]] = None,
 ) -> int:
     """Upload a file at specified file path to a url.
 
@@ -230,6 +241,7 @@ def upload_file(
             should be included in the signing procedure.
         callback_func: optional function used to receive: bytes_uploaded, total_bytes
             for example: def callback(bytes_uploaded: int, total_bytes: int) -> None
+        headers_factory: optional function to inject headers
 
     Returns:
         The total number of uploaded bytes.
@@ -257,6 +269,7 @@ def upload_file(
             pool=pool,
             md5=md5,
             callback_func=callback_func,
+            headers_factory=headers_factory,
         )
 
     return size
@@ -311,6 +324,7 @@ def upload_fileobj(
     pool: Optional[urllib3.PoolManager] = None,
     md5: Optional[bytes] = None,
     callback_func: Optional[Callable[[int, int], None]] = None,
+    headers_factory: Optional[Callable[[], Dict[str, str]]] = None,
 ) -> int:
     """Upload a file object to a url.
 
@@ -331,6 +345,7 @@ def upload_fileobj(
             should be included in the signing procedure.
         callback_func: optional function used to receive: bytes_uploaded, total_bytes
             for example: def callback(bytes_uploaded: int, total_bytes: int) -> None
+        headers_factory: optional function to inject headers
 
     Returns:
         The total number of uploaded bytes.
@@ -384,6 +399,8 @@ def upload_fileobj(
     }
     if md5 is not None:
         headers["Content-MD5"] = base64.b64encode(md5).decode()
+    if headers_factory is not None:
+        headers.update(headers_factory())
     response = pool.request(
         "PUT",
         url,